techdebt(parallel): split skeleton.rs into orchestration / legacy / diff / tests submodules
`crates/tsz-core/src/parallel/skeleton.rs` was 3187 lines — over the §19
2000-line cap — and interleaved skeleton production, legacy-merge parity
assertions, incremental-diff helpers, and a 1.2k-line test module. The mixing
made it hard to tell which retained branches were carrying current parallel
checking vs preserved-for-parity legacy paths.

Converts `skeleton.rs` to a `skeleton/` module directory:

- `skeleton/mod.rs` (1598 lines) — `FileSkeleton`/`SkeletonIndex` types,
  `extract_skeleton`, `reduce_skeletons`, and the `SkeletonIndex` lookup +
  `build_*` projection methods. Reads as forward data flow from `BindResult`
  → file skeleton → reduced index.
- `skeleton/legacy_validation.rs` — the five `validate_*_against_legacy`
  /`validate_against_merged` debug-only parity assertions, gathered in one
  named submodule with each method commented with the test(s) that pin it.
  No-op in release builds.
- `skeleton/diff.rs` — `SkeletonDiff` + `diff_skeletons`. Re-exported from
  `skeleton/mod.rs` so external callers (LSP / `parallel_tests`) keep their
  existing import paths.
- `skeleton/tests.rs` — the `#[cfg(test)] mod tests` block, kept verbatim.

Behavior-preserving: every moved item retains its signature and body, and
`parallel/mod.rs::pub use skeleton::*` continues to surface the same public
API. Verified via `cargo test -p tsz-core --lib parallel::` (283 passed).

Closes #9426

AgentName: claude-opus-4-7-1m
